requirements.txt
hydra-core==1.3.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
numpy==1.24.4
omegaconf==2.3.0
scikit-learn==1.0.2
scipy==1.10.1
six==1.16.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchnet==0.0.4
torchvision==0.9.1+cu111
tornado==6.4.1